Take the default title and description from commit when creating merge request.
When creating a new merge request it will now only set the default title and description if both the source branch and target branch is selected.

If the merge request consists of a single commit, the commit message will be used to set the default title and description. If the merge request consists of more than one commit it will fallback to the current behavior of using the name of the source branch for the title and leave the description blank.